backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
numpy==1.26.0
tensorflow==2.14.0
gunicorn==21.2.0
scikit-learn==1.3.1
opencv-python==4.8.1.78